Fix compile warning.
git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@6151 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_mmo_champion.cpp
@@ -339,7 +339,7 @@ static bool parse_item_name( item_t&            item,
   if ( pos != std::string::npos ) s.erase( pos );
 
   // The MMO-Champion names often have numbers embedded in the name.....
-  for ( int i=0; i < s.size(); i++ )
+  for ( int i=0; i < (int) s.size(); i++ )
   {
     if ( isdigit( s[ i ] ) )
     {
